+++ app/controllers/homepage_controller.rb
@@ -4,7 +4,6 @@ class HomepageController < ApplicationController
     @background = ''
     @ranks_west = Team.rankings_conf('W')
     @ranks_east = Team.rankings_conf('E')
-    @r_cnt = 0
     @todays_best = []
     @games = [[],[]]
     @results = [[],[]]
@@ -26,9 +25,6 @@ class HomepageController < ApplicationController
       @results[1].each {|x| game_list << x.boxscore_id}
     end
     @todays_best = Gamestat.topGameScores(game_list)
-    # @todays_best.each do |p|
-    #   p.bg = @logos[Team.getTeamName(Team.getTeamId(p.t_abbr))]
-    # end
 
   end