+++ app/models/team.rb
@@ -8,7 +8,7 @@ class Team < ActiveRecord::Base
   # @param s [String] #Team Abbreviation
   # @return [Integer] # Team ID
   def self.getTeamId(s="")
-    t = Team.find_by("t_abbr = ?", s.to_s.upcase)
+    t = Team.find_by("abbr = ?", s.to_s.upcase)
     return (!t.nil? ? t.id : nil)
   end
 
@@ -17,12 +17,12 @@ class Team < ActiveRecord::Base
   # @return [String] # Team Abbr
   def self.getTeamAbbr(i)
     t = Team.find(i)
-    return (!t.nil? ? t.t_abbr : nil)
+    return (!t.nil? ? t.abbr : nil)
   end
 
   def self.getTeamName(id)
     if(!id.nil?)
-      return Team.find(id).t_name
+      return Team.find(id).name
     end
     return nil
   end
@@ -37,10 +37,10 @@ class Team < ActiveRecord::Base
     t_list = []
     t_rank = []
     teams.each { |t| t_list << t.id }
-    @records = Game.where(team_id: t_list, season_type: 2).where.not(:boxscore_id => 0).order("gdate desc").group(:team_id)
+    @records = Game.where(team_id: t_list, season_type: 2).where.not(:boxscore_id => 0).order("datetime desc").group(:team_id)
     @records.each do |record|
       w_pct = record.wins.to_f / (record.wins + record.losses)
-      t_rank << [w_pct.round(2), Team.find(record.team_id).t_name, record]
+      t_rank << [w_pct.round(2), Team.find(record.team_id).name, record]
     end
     t_rank.sort!.reverse!
   end