+++ db/SeedDB.rb
@@ -14,7 +14,7 @@ class SeedDB
 
   # Use EspnScrape to collect live data from the internet
   def seed_from_web
-    @file = SAVE_TO_FILE ? File.open('db/'+STORED_SEEDS, "w") : nil # Destination file for saving
+    @file = File.open('db/'+STORED_SEEDS, "w") if SAVE_TO_FILE # Destination file for saving
     seed_teams                # EspnScrape::NbaTeamList => Teams
     Team.all.each do |team|
       process_schedules(team) # EspnScrape::NbaSchedule => Games
@@ -27,9 +27,9 @@ class SeedDB
   # Populate Teams Table with NBA League data
   def seed_teams
     print "\nSeeding Teams..."
-    if(Team.all.count == 0)                 # Skip already populated data
-      fl_a = EspnScrape.teamList.to_hashes  # Field List Array => [Dictionary]
-      Team.create(fl_a)                     # Create all Teams
+    if(Team.all.count == 0)                     # Skip already populated data
+      fl_a = EspnScrape.teamList(:to_hashes)[]  # Field List Array => [Dictionary]
+      Team.create(fl_a)                         # Create all Teams
       @file.puts "Team.create(#{fl_a})" if SAVE_TO_FILE  # Save Create command
     end
     puts "#{Team.all.count}...Done."                            # Confirm Team count
@@ -38,11 +38,11 @@ class SeedDB
 
   # Populate Games Table with Team Schedule Data
   def process_schedules(team)
-    print "-- #{team.t_name} Games..."
+    print "-- #{team.name} Games..."
     SEASON_TYPES.each { |season_type|                         # Process configured season types
-      t_schedule = EspnScrape.schedule(team.t_abbr, season_type)
+      t_schedule = EspnScrape.schedule(team.abbr, season_type, :to_hashes)
       # Collect boxscore ids for Gamestats population
-      @completedGames += t_schedule.pastGames.to_hashes.map { |pg| pg[:boxscore_id] }
+      @completedGames += t_schedule.pastGames[].map { |pg| pg[:boxscore_id] }
 
       # Skip already populated data
       if(Game.where("team_id = ? AND season_type = ?", team.id, season_type).size == 0)
@@ -56,10 +56,10 @@ class SeedDB
   # Populate Players Table with Team Roster Data
   def process_roster(team)
     ## Process Team Roster
-    print "-- #{team.t_name} Players..."
+    print "-- #{team.name} Players..."
     if(Player.where("team_id = ?", team.id).size == 0)        # Skip already populated data
-      roster = EspnScrape.roster(team.t_abbr)
-      fl_a = Player.refreshRoster(team, roster.players)
+      roster = EspnScrape.roster(team.abbr, :to_hashes)
+      fl_a = Player.refreshRoster(team, roster.players[])
       @file.puts "Player.create(#{fl_a})" if SAVE_TO_FILE  # Save Create command
     end
     puts "%i...Done." % [Player.where("team_id = ?", team.id).size] # Confirm Player count
@@ -75,19 +75,19 @@ class SeedDB
       print "-- Gamestats : #{idx} / #{out_of} [#{((idx/out_of.to_f)*100).round}\%]    \r"
       $stdout.flush
       if Gamestat.find_by("boxscore_id = ?", boxscore_id).nil?  # Skip already processed Games
-        bs = EspnScrape.boxscore(boxscore_id)
+        bs = EspnScrape.boxscore(boxscore_id, :to_hashes)
 
-        fl_a_home = bs.homePlayers.to_hashes
-        fl_a_away = bs.awayPlayers.to_hashes
+        fl_a_home = bs.homePlayers[]
+        fl_a_away = bs.awayPlayers[]
 
         # Set Foreign Keys - Used to simplify site navigation
         fl_a_home.each do |fl|
           fl = Player.setForeignKeys(fl, boxscore_id, bs, true, @file)
-          fl[:game_num] = Game.getGameNum(fl[:t_abbr], boxscore_id)
+          fl[:game_num] = Game.getGameNum(fl[:abbr], boxscore_id)
         end
         fl_a_away.each do |fl|
           fl = Player.setForeignKeys(fl, boxscore_id, bs, false, @file)
-          fl[:game_num] = Game.getGameNum(fl[:t_abbr], boxscore_id)
+          fl[:game_num] = Game.getGameNum(fl[:abbr], boxscore_id)
         end
       end
       Gamestat.create(fl_a_away)