+++ db/migrate/20151120235554_create_gamestats.rb
@@ -3,14 +3,14 @@ class CreateGamestats < ActiveRecord::Migration
     create_table :gamestats do |t|
       t.references :player,   index: true
       t.integer :boxscore_id, index: true
-      t.string  :t_abbr,      index: true
-      t.integer :opp_id,      index: true
-      t.string  :opp_abbr,    index: true
+      t.string  :abbr,        index: true
+      t.integer :opponent_id, index: true
+      t.string  :opponent,    index: true
       t.integer :game_num
-      t.string  :p_name
-      t.integer :p_eid
-      t.string  :pos
-      t.integer :min
+      t.string  :name
+      t.integer :eid
+      t.string  :position
+      t.integer :minutes
       t.integer :fgm
       t.integer :fga
       t.integer :tpm
@@ -19,14 +19,14 @@ class CreateGamestats < ActiveRecord::Migration
       t.integer :fta
       t.integer :oreb
       t.integer :dreb
-      t.integer :reb
-      t.integer :ast
-      t.integer :stl
-      t.integer :blk
-      t.integer :tos
-      t.integer :pf
-      t.integer :plus
-      t.integer :pts
+      t.integer :rebounds
+      t.integer :assists
+      t.integer :steals
+      t.integer :blocks
+      t.integer :turnovers
+      t.integer :fouls
+      t.integer :plusminus
+      t.integer :points
       t.boolean :starter
 
       t.timestamps null: false