+++ db/migrate/20151125075938_create_players.rb
@@ -2,15 +2,14 @@ class CreatePlayers < ActiveRecord::Migration
   def change
     create_table :players do |t|
       t.references :team, index: true
-      # t.references :game, index: true
-      t.string  :t_abbr, index: true
-      t.integer :p_eid,  index: true
-      t.string  :p_name
-      t.integer :p_num
-      t.string  :pos
+      t.string  :abbr, index: true
+      t.integer :eid,  index: true
+      t.string  :name
+      t.integer :jersey
+      t.string  :position
       t.integer :age
-      t.integer :h_ft
-      t.integer :h_in
+      t.integer :height_ft
+      t.integer :height_in
       t.integer :weight
       t.string  :college
       t.integer :salary